+++ libre_pvz/src/main.rs
@@ -19,9 +19,11 @@
 // use std::path::{Path, PathBuf};
 use bevy::prelude::*;
 use bevy_egui::EguiPlugin;
+use bevy_inspector_egui::WorldInspectorPlugin;
 use libre_pvz::animation::AnimationPlugin;
 use libre_pvz::core::kinematics::KinematicsPlugin;
 use libre_pvz::diagnostics::BoundingBoxPlugin;
+use libre_pvz::plant::peashooter::PeashooterPlugin;
 // use libre_pvz::scene::almanac::AlmanacPlugin;
 use libre_pvz::scene::lawn::LawnPlugin;
 use libre_pvz_resources::ResourcesPlugin;
@@ -37,10 +39,12 @@ fn main() {
         // .insert_resource(AlmanacPlugin::window_descriptor())
         .add_plugins(DefaultPlugins)
         .add_plugin(EguiPlugin)
+        .add_plugin(WorldInspectorPlugin::new())
         .add_plugin(BoundingBoxPlugin)
         .add_plugin(AnimationPlugin)
         .add_plugin(ResourcesPlugin)
         .add_plugin(KinematicsPlugin)
+        .add_plugin(PeashooterPlugin)
         .add_plugin(LawnPlugin)
         // .add_plugin(AlmanacPlugin::new(anim_name))
         .add_startup_system(setup_camera)