+++ libre_pvz/src/plant.rs
@@ -16,16 +16,6 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-//! librePvZ: game logic implementation.
+//! Plant-related things.
 
-#![warn(missing_docs)]
-#![warn(missing_debug_implementations)]
-
-pub use libre_pvz_resources as resources;
-pub use libre_pvz_animation as animation;
-
-pub mod core;
-pub mod scene;
-pub mod diagnostics;
-
-pub mod plant;
+pub mod peashooter;