+++ libre_pvz/CHANGELOG.md
@@ -15,4 +15,5 @@ This project adheres to [Cargo's Semantic Versioning](https://doc.rust-lang.org/
 ### Lawn Scene
 
 - show day lawn, set up a grid system on the lawn
+- defined Peashooter using the Model resource
 - show a hard-coded Repeater on the lawn